automatically backed up by learn

+++ db/migrate/20171023223731_create_table_users.rb
@@ -3,6 +3,6 @@ class CreateTableUsers < ActiveRecord::Migration
     create_table :users do |u|
       u.string :username
       u.string :password
-      u.float :balance
+      u.decimal :balance
   end
 end
